Rename and reorganize folder structure and project name
