+++ README.md
@@ -1,2 +1,2 @@
-# Spring Boot Hello World
-This project is an introductory project for building an understanding of the Spring Boot framework as well as setting up simple REST APIs with microservice architecture.
+# Accounts API
+A microservice application for managing account information for a bank.
