+++ src/main/java/com/hellobank/AccountsAPI.java
@@ -1,14 +1,14 @@
-package com.example.springboothelloworld;
+package com.hellobank;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SpringBootHelloWorldApplication {
+public class AccountsAPI {
 
 	public static void main(String[] args) {
 
-		SpringApplication.run(SpringBootHelloWorldApplication.class, args);
+		SpringApplication.run(AccountsAPI.class, args);
 
 	}
