+++ src/main/java/com/hellobank/account/controller/domain/AccountControllerResponse.java
@@ -1,4 +1,4 @@
-package com.example.springboothelloworld.account.controller.domain;
+package com.hellobank.account.controller.domain;
 
 public interface AccountControllerResponse {
     // Used for abstracting generalizing REST API responses.
