+++ src/main/java/com/hellobank/account/domain/Account.java
@@ -1,4 +1,4 @@
-package com.example.springboothelloworld.account.domain;
+package com.hellobank.account.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
